Move file contents into graph/special.py
